+++ src/components/Details.js
@@ -2,11 +2,11 @@ import React from "react";
 import { useFetch } from "../App";
 
 export function Details() {
-    const { data, details, openMenu, closeMenu } = useFetch();
+  const { data, details, openMenu, closeMenu } = useFetch();
   return (
     <div className="details-2">
       <div className="photo-description">
-        <img  alt="" />
+        <img alt="" />
         <div className="description-2"></div>
       </div>
