+++ src/components/UseCards.js
@@ -1,21 +1,36 @@
-import React, { useEffect, useState } from 'react'
+import React, { useEffect, useState } from "react";
 
 export const UseCards = () => {
-    const [data,setData]=useState();
-    const [nutrition,setNutrition]=useState();
- useEffect(()=>{
-     fetch(`https://fruits-develhope.herokuapp.com/api`)
-     .then(response=>{
-        return response.json()
-     })
-     .then(json=>{
-         setData(json)
-     })
-     
- },[])
+  const [data, setData] = useState();
+  const [nutrition, setNutrition] = useState();
+  const [details, setDetails] = useState(false);
+  const [element, setElement] = useState();
+  useEffect(() => {
+    fetch(`https://fruits-develhope.herokuapp.com/api`)
+      .then((response) => {
+        return response.json();
+      })
+      .then((json) => {
+        setData(json);
+      });
+  }, []);
+
+  const openMenu = function OpenDetails(data) {
+    setElement(data);
+    console.log(data);
+    setDetails(true);
+    console.log(details);
+  };
+  const closeMenu = function CloseDetails() {
+    setDetails(false);
+  };
 
   return {
     nutrition,
-      data,
-  }
-}
+    data,
+    openMenu,
+    closeMenu,
+    details,
+    element,
+  };
+};
